+++ requirements-dev.txt
@@ -6,3 +6,4 @@ py-geth>=0.5.0
 # Until pyethereum > 1.3.6 is released.
 https://github.com/ethereum/pyethereum/tarball/b06829e56e2a3de5276077a611ed8813b6cf5e74
 secp256k1>=0.12.0
+rlp==0.4.6
