+++ web3/utils/encoding.py
@@ -4,40 +4,77 @@ import sys
 import binascii
 import json
 
-from . import utils
-
-from rlp.utils import (
-    decode_hex as _decode_hex,
-    encode_hex as _encode_hex,
+from .types import (
+    is_bytes,
+    is_text,
+    is_string,
+    is_number,
+    is_integer,
+    is_boolean,
+    is_object,
+)
+from .formatting import (
+    remove_0x_prefix,
+    add_0x_prefix,
 )
 
 
-def toHex(val):
+if sys.version_info.major == 2:
+    def decode_hex(value):
+        if isinstance(value, bytearray):
+            value = str(value)
+        if not is_string(value):
+            raise TypeError('Value must be an instance of str or unicode')
+        return remove_0x_prefix(value).decode('hex')
+
+    def encode_hex(value):
+        if isinstance(value, bytearray):
+            value = str(value)
+        if not is_string(value):
+            raise TypeError('Value must be an instance of str or unicode')
+        return add_0x_prefix(value.encode('hex'))
+else:
+    def decode_hex(s):
+        if isinstance(s, str):
+            return bytes.fromhex(remove_0x_prefix(s))
+        if isinstance(s, bytes):
+            return binascii.unhexlify(remove_0x_prefix(s))
+        raise TypeError('Value must be an instance of str or bytes')
+
+    def encode_hex(b):
+        if isinstance(b, str):
+            b = bytes(b, 'utf-8')
+        if isinstance(b, bytes):
+            return add_0x_prefix(binascii.hexlify(b))
+        raise TypeError('Value must be an instance of str or bytes')
+
+
+def to_hex(value):
     """
     Auto converts any given value into it's hex representation.
     """
-    if utils.isBoolean(val):
-        return "0x1" if val else "0x0"
+    if is_boolean(value):
+        return "0x1" if value else "0x0"
 
-    if utils.isObject(val):
-        return fromUtf8(json.dumps(val))
+    if is_object(value):
+        return encode_hex(json.dumps(value))
 
-    if utils.isString(val):
-        if val.startswith("-0x"):
-            return fromDecimal(val)
-        elif val.startswith("0x"):
-            return val
-        elif not utils.isNumber(val):
-            return fromAscii(val)
+    if is_string(value):
+        if value.startswith("-0x"):
+            return fromDecimal(value)
+        elif value.startswith("0x"):
+            return value
+        elif not is_number(value):
+            return fromAscii(value)
 
-    return fromDecimal(val)
+    return from_decimal(val)
 
 
-def toDecimal(value):
+def to_decimal(value):
     """
     Converts value to it's decimal representation in string
     """
-    if utils.isString(value):
+    if is_string(value):
         if value.startswith("0x") or value.startswith("-0x"):
             value = int(value, 16)
         else:
@@ -80,13 +117,6 @@ def toAscii(hex):
     return binascii.unhexlify(hex).decode("ascii")
 
 
-def fromUtf8(str):
-    """
-    Should be called to get hex representation (prefixed by 0x) of utf8 string
-    """
-    return "0x" + binascii.hexlify(str.encode("utf8")).decode("utf8")
-
-
 def fromAscii(obj):
     """
     Should be called to get hex representation (prefixed by 0x) of ascii string
@@ -99,129 +129,3 @@ def abiToJson(obj):
     Converts abi string to python object
     """
     return json.loads(obj)
-
-
-if sys.version_info.major == 2:
-    integer_types = (int, long)  # NOQA
-    bytes_types = (bytes, bytearray)
-    text_types = (unicode,)  # NOQA
-    string_types = (basestring, bytearray)  # NOQA
-else:
-    integer_types = (int,)
-    bytes_types = (bytes, bytearray)
-    text_types = (str,)
-    string_types = (bytes, str, bytearray)
-
-
-def is_integer(value):
-    return isinstance(value, integer_types)
-
-
-def is_bytes(value):
-    return isinstance(value, bytes_types)
-
-
-def is_text(value):
-    return isinstance(value, text_types)
-
-
-def is_string(value):
-    return isinstance(value, string_types)
-
-
-if sys.version_info.major == 2:
-    def force_bytes(value):
-        if is_bytes(value):
-            return str(value)
-        elif is_text(value):
-            return value.encode('latin1')
-        else:
-            raise TypeError("Unsupported type: {0}".format(type(value)))
-
-    def force_text(value):
-        if is_text(value):
-            return value
-        elif is_bytes(value):
-            return unicode(force_bytes(value), 'latin1')  # NOQA
-        else:
-            raise TypeError("Unsupported type: {0}".format(type(value)))
-else:
-    def force_bytes(value):
-        if is_bytes(value):
-            return bytes(value)
-        elif is_text(value):
-            return bytes(value, 'latin1')
-        else:
-            raise TypeError("Unsupported type: {0}".format(type(value)))
-
-    def force_text(value):
-        if isinstance(value, text_types):
-            return value
-        elif isinstance(value, bytes_types):
-            return str(value, 'latin1')
-        else:
-            raise TypeError("Unsupported type: {0}".format(type(value)))
-
-
-def force_obj_to_bytes(obj):
-    if is_string(obj):
-        return force_bytes(obj)
-    elif isinstance(obj, dict):
-        return {
-            k: force_obj_to_bytes(v) for k, v in obj.items()
-        }
-    elif isinstance(obj, (list, tuple)):
-        return type(obj)(force_obj_to_bytes(v) for v in obj)
-    else:
-        return obj
-
-
-def force_obj_to_text(obj):
-    if is_string(obj):
-        return force_text(obj)
-    elif isinstance(obj, dict):
-        return {
-            k: force_obj_to_text(v) for k, v in obj.items()
-        }
-    elif isinstance(obj, (list, tuple)):
-        return type(obj)(force_obj_to_text(v) for v in obj)
-    else:
-        return obj
-
-
-def coerce_args_to_bytes(fn):
-    @functools.wraps(fn)
-    def inner(*args, **kwargs):
-        bytes_args = force_obj_to_bytes(args)
-        bytes_kwargs = force_obj_to_bytes(kwargs)
-        return fn(*bytes_args, **bytes_kwargs)
-    return inner
-
-
-def coerce_return_to_bytes(fn):
-    @functools.wraps(fn)
-    def inner(*args, **kwargs):
-        return force_obj_to_bytes(fn(*args, **kwargs))
-    return inner
-
-
-@coerce_args_to_bytes
-def remove_0x_prefix(value):
-    if value.startswith(b'0x'):
-        return value[2:]
-    return value
-
-
-@coerce_args_to_bytes
-def add_0x_prefix(value):
-    return b'0x' + remove_0x_prefix(value)
-
-
-@coerce_args_to_bytes
-def decode_hex(value):
-    return _decode_hex(remove_0x_prefix(value))
-
-
-@coerce_args_to_bytes
-def encode_hex(value):
-    return add_0x_prefix(_encode_hex(value))
